Group Actions in ActionStack instead of merging
It was pointed out to me, that you sometimes want to have multiple edits
in different locations grouped as a single undo action (for example
'replace all'), but with my previous system, a single undo-able action
could only be one contiguous modification of the text. I could have
merged those still, but the merging itself smelled dangerous and you
would have different code for insertions and deletions and so on.
Also later I want to provide a Buffer::startUndoTransation and
Buffer::endUntoTransaction() api, so decided to have the ActionStack be
able to group Actions as one. I think this is somewhat inefficient for
large groupings, but I want to see it be too slow, before I think about
that more.